refactor(UNT-T15674): refactor styles file path move into directory
