+++ src/components/PhotoGallery/components/styles/NetworkImageStyle.ts
