+++ src/components/PhotoGallery/components/styles/NetworkLoaderStyle.ts
