+++ src/components/PhotoGallery/components/styles/index.ts
@@ -1,0 +1,11 @@
+import ListItemStyle from './ListItemStyle';
+import PhotoModalStyle from './PhotoModalStyle';
+import NetworkImageStyle from './NetworkImageStyle';
+import NetworkLoaderStyle from './NetworkLoaderStyle';
+
+export {
+  ListItemStyle,
+  PhotoModalStyle,
+  NetworkImageStyle,
+  NetworkLoaderStyle,
+};
